refactor: remove client role assignment (#18)
Remove role assignment for current client, as role "Storage Blob Data Contributor" is not required to create Containers when shared access key is disabled.

This will trigger the destruction of resource "azurerm_role_assignment, however this will not affect the functionality of this module. The role assignment can be recreated in the calling module if needed.